+++ tapeserver/castor/tape/tapeserver/file/OsmFileReader.hpp
@@ -17,20 +17,19 @@
 
 #pragma once
 
-#include <memory>
 
-#include "castor/tape/tapeserver/file/OsmFileReader.hpp"
+#include "castor/tape/tapeserver/file/CpioFileHeaderStructure.hpp"
+#include "tapeserver/castor/tape/tapeserver/file/OsmFileStructure.hpp"
 #include "castor/tape/tapeserver/file/FileReader.hpp"
+
+#include <memory>
 
 namespace castor {
 namespace tape {
 namespace tapeFile {
-
-class UHL1;
-
+  
 class OsmFileReader : public FileReader {
 public:
-  CTA_GENERATE_EXCEPTION_CLASS(NotImplemented);
   /**
     * Constructor of the FileReader. It will bind itself to an existing read session
     * and position the tape right at the beginning of the file
@@ -46,10 +45,31 @@ public:
   ~OsmFileReader() override = default;
 
   size_t readNextDataBlock(void *data, const size_t size) override;
+  
+  /**
+   *
+   */
+  inline const CPIO& getCPIOHeader() const {
+    return m_cpioHeader;
+  }
+
 
 private:
+  const size_t PAYLOAD_BOLCK_SIZE = 262144;
+  /*
+   * CPIO file
+   */
+  CPIO m_cpioHeader;
+  uint64_t m_ui64CPIODataSize = 0;
+
+  
   void positionByFseq(const cta::RetrieveJob &fileToRecall) override;
   void positionByBlockID(const cta::RetrieveJob &fileToRecall) override;
+  
+  void moveToFirstFile();
+  void moveReaderByFSeqDelta(const int64_t fSeq_delta);
+  void useBlockID(const cta::RetrieveJob &fileToRecall);
+  void setBlockSize(size_t uiBlockSize);
 };
 
 }  // namespace tapeFile